feat(discord-bot): webhookに新しいコマンドと機能を追加
